+++ src/utils/icons.tsx
@@ -19,7 +19,7 @@ import {
 } from 'react-icons/fa';
 import { AiFillFilePdf, AiFillFileImage, AiFillFileWord } from 'react-icons/ai';
 import { IoIosSearch, IoIosEye, IoIosEyeOff, IoIosAddCircle, IoIosCloseCircleOutline, IoIosArrowBack, IoMdArrowDropright } from 'react-icons/io';
-import { MdFormatListBulleted, MdLocationOn, MdOutlineCreateNewFolder, MdTitle } from 'react-icons/md';
+import { MdContentCopy, MdFormatListBulleted, MdLocationOn, MdOutlineCreateNewFolder, MdTitle } from 'react-icons/md';
 import { CgChevronDoubleDown, CgChevronDoubleUp, CgSortAz } from 'react-icons/cg';
 import { LuListChecks } from 'react-icons/lu';
 
@@ -156,6 +156,7 @@ export {
     CgSortAz,
     MdOutlineCreateNewFolder,
     MdFormatListBulleted,
+    MdContentCopy,
     MdLocationOn,
     MdTitle,
     LuListChecks,